refactor: CircleMembership を独立ディレクトリに分離 (#575)
CircleMembership の Domain 層・Infrastructure 層ファイルを
circle/ から circle-membership/ 専用ディレクトリに移動し、
Aggregate Root 単位のディレクトリ構成に統一。
