+++ server/application/circle/circle-membership-service.ts
@@ -1,6 +1,6 @@
 import { userId, type CircleId, type UserId } from "@/server/domain/common/ids";
-import type { CircleMembershipRepository } from "@/server/domain/models/circle/circle-membership-repository";
-import type { CircleMembership } from "@/server/domain/models/circle/circle-membership";
+import type { CircleMembershipRepository } from "@/server/domain/models/circle-membership/circle-membership-repository";
+import type { CircleMembership } from "@/server/domain/models/circle-membership/circle-membership";
 import type { CircleRepository } from "@/server/domain/models/circle/circle-repository";
 import type { createAccessService } from "@/server/application/authz/access-service";
 import type {
