+++ server/application/circle/circle-service.ts
@@ -6,7 +6,7 @@ import {
 import { userId } from "@/server/domain/common/ids";
 import type { CircleId } from "@/server/domain/common/ids";
 import type { CircleRepository } from "@/server/domain/models/circle/circle-repository";
-import type { CircleMembershipRepository } from "@/server/domain/models/circle/circle-membership-repository";
+import type { CircleMembershipRepository } from "@/server/domain/models/circle-membership/circle-membership-repository";
 import type { createAccessService } from "@/server/application/authz/access-service";
 import type {
   Repositories,
