+++ server/domain/models/circle-membership/circle-membership-repository.ts
@@ -1,5 +1,5 @@
 import type { CircleId, UserId } from "@/server/domain/common/ids";
-import type { CircleMembership } from "@/server/domain/models/circle/circle-membership";
+import type { CircleMembership } from "@/server/domain/models/circle-membership/circle-membership";
 import type { CircleRole } from "@/server/domain/services/authz/roles";
 
 export type CircleMembershipRepository = {
