+++ server/infrastructure/mappers/circle-membership-mapper.ts
@@ -1,5 +1,5 @@
 import { circleId, userId } from "@/server/domain/common/ids";
-import type { CircleMembership } from "@/server/domain/models/circle/circle-membership";
+import type { CircleMembership } from "@/server/domain/models/circle-membership/circle-membership";
 import type { CircleRole } from "@/server/domain/services/authz/roles";
 import type { CircleRole as PrismaCircleRole } from "@/generated/prisma/enums";
