+++ server/infrastructure/transaction/prisma-unit-of-work.ts
@@ -1,7 +1,7 @@
 import type { UnitOfWork } from "@/server/application/common/unit-of-work";
 import { prisma } from "@/server/infrastructure/db";
 import { createPrismaCircleRepository } from "@/server/infrastructure/repository/circle/prisma-circle-repository";
-import { createPrismaCircleMembershipRepository } from "@/server/infrastructure/repository/circle/prisma-circle-membership-repository";
+import { createPrismaCircleMembershipRepository } from "@/server/infrastructure/repository/circle-membership/prisma-circle-membership-repository";
 import { createPrismaCircleSessionRepository } from "@/server/infrastructure/repository/circle-session/prisma-circle-session-repository";
 import { createPrismaCircleSessionMembershipRepository } from "@/server/infrastructure/repository/circle-session/prisma-circle-session-membership-repository";
 import { createPrismaMatchRepository } from "@/server/infrastructure/repository/match/prisma-match-repository";
